backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
python-dotenv==1.1.1
sqlalchemy==2.0.43
asyncpg==0.30.0
greenlet==3.2.4
pandas==2.3.2
httpx==0.28.1
openpyxl==3.1.2
aiofiles==23.2.1
email-validator==2.2.0
jinja2==3.1.5